vehicle-building-detection/code/inference/requirements.txt
Pillow
tensorflow-gpu==1.4.0
numpy
tqdm
